third_party/WebKit/LayoutTests/fast/css/usecounter-negative-background-size-shorthand.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let NegativeBackgroundSize = 2402; // From web_feature.mojom

    let isCounted = () => internals.isUseCounted(document, NegativeBackgroundSize);
    var div = document.createElement('div');

    div.style = 'background: none, red none 50% 50% / 1px 2px round space local padding-box content-box;';
    div.style = '-webkit-mask: none -1px -2px, none -3px -4px;';
    assert_false(isCounted(),
                 '-webkit-mask should not be counted');

    div.style = 'background: none, red none 50% 50% / -1px 2px round space local padding-box content-box;';
    assert_true(isCounted(),
                'background should be counted');
}, 'Negative size is use counted for background-size in shorthand');
</script>
